Move local Razor helpers out of SharedUtilities
