+++ src/Razor/src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/Utilities/Delta.cs
