+++ src/Razor/src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/VoidResult.cs
