+++ src/Razor/src/Razor/src/Microsoft.VisualStudio.LanguageServices.Razor/Snippets/SnippetCache.cs
@@ -5,7 +5,6 @@ using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.ComponentModel.Composition;
 using System.Linq;
-using Microsoft.AspNetCore.Razor;
 using Microsoft.AspNetCore.Razor.PooledObjects;
 using static Microsoft.VisualStudio.Razor.Snippets.XmlSnippetParser;
