+++ src/Razor/src/Shared/Microsoft.AspNetCore.Razor.SharedUtilities/Threading/SpecializedTasks.cs
