+++ src/Razor/src/Shared/Microsoft.AspNetCore.Razor.SharedUtilities/Threading/TaskExtensions.cs
